requirements.txt


Django==5.0.3
djangorestframework==3.15.2
drf-yasg==1.21.7
django-environ==0.11.2
